profile group updated

+++ client/src/main/java/com/whisper/client/presentation/controllers/profileGroupController.java
@@ -4,8 +4,11 @@ import com.whisper.client.HelloApplication;
 import com.whisper.client.MyApp;
 import com.whisper.client.business.services.ChattingService;
 import com.whisper.client.business.services.ContactService;
+import com.whisper.client.business.services.GroupService;
+import com.whisper.client.presentation.services.DialogueManager;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
+import javafx.embed.swing.SwingFXUtils;
 import javafx.event.ActionEvent;
 import javafx.event.Event;
 import javafx.fxml.FXMLLoader;
@@ -14,17 +17,23 @@ import javafx.scene.control.*;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.HBox;
+import javafx.stage.FileChooser;
+import javafx.stage.Stage;
 import org.example.entities.RoomChat;
 import org.example.entities.User;
 import org.example.serverinterfaces.ChatServiceInt;
+import org.example.serverinterfaces.EditGroupServiceInt;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
+import javax.imageio.ImageIO;
+import javax.sql.rowset.serial.SerialBlob;
+import java.io.*;
 import java.net.URL;
 import java.rmi.NotBoundException;
 import java.rmi.RemoteException;
 import java.rmi.registry.LocateRegistry;
 import java.rmi.registry.Registry;
+import java.sql.Blob;
+import java.sql.SQLException;
 import java.util.List;
 import java.util.ResourceBundle;
 
@@ -59,16 +68,45 @@ public class profileGroupController implements Initializable {
     HBox hBox;
     RoomChat roomChat;
 
-//    List<User> users = new ArrayList<>(users());
-
+    DialogueManager dialogueManager = DialogueManager.getInstance();
+    GroupService groupService = new GroupService();
 
     @javafx.fxml.FXML
     public void onProfileClicked(Event event) {
+        FileChooser fileChooser = new FileChooser();
+        fileChooser.getExtensionFilters().addAll(
+                new FileChooser.ExtensionFilter("Image Files", "*.png", "*.jpg", "*.jpeg")
+        );
+        File file = fileChooser.showOpenDialog(new Stage());
+        if (file != null) {
+            Image image = new Image(file.toURI().toString());
+            groupImage.setImage(image);
+        }
+    }
 
+    private byte[] imageViewToByteArray(ImageView imageView) {
+        Image image = imageView.getImage();
+        ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
+        try {
+            ImageIO.write(SwingFXUtils.fromFXImage(image, null), "png", byteArrayOutputStream);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return byteArrayOutputStream.toByteArray();
     }
 
     @javafx.fxml.FXML
     public void onSaveChangesClicked(ActionEvent actionEvent) {
+        roomChat.setGroupName(groupName.getText());
+        roomChat.setDescription(groupDescription.getText());
+        byte[] imageByteArray = imageViewToByteArray(groupImage);
+        try {
+            Blob imageBlob = new SerialBlob(imageByteArray);
+            roomChat.setPhotoBlob(imageBlob); // Set the photo blob with the Blob object
+        } catch (SQLException e) {
+            e.printStackTrace(); // Handle the exception appropriately
+        }
+        groupService.editGroup(roomChat);
     }
 
     @Override
@@ -79,7 +117,6 @@ public class profileGroupController implements Initializable {
         groupMembers = chattingService.getGroupMembers(5);
         System.out.println(groupMembers.size());
         roomChat = chattingService.getRoomChatByID(5);
-
 
         for (User user: groupMembers){
             FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("views/groupMemberView.fxml"));
@@ -94,17 +131,12 @@ public class profileGroupController implements Initializable {
         }
         groupMembersList.setItems(boxes);
 
-//        System.out.println(otherContacts.size());
-//        otherContacts.removeIf(contact -> groupMembers.contains(contact));
-//        otherContacts.removeAll(groupMembers);
-//
-//        // Remove elements from otherContacts that exist in groupMembers
-//        otherContacts.removeAll(groupMembers);
-
         for (int i=0; i<groupMembers.size(); i++){
             for (int j=0; j<otherContacts.size(); j++){
                 if (groupMembers.get(i).getUserId() == otherContacts.get(j).getUserId())
                     otherContacts.remove(j);
+                if (groupMembers.get(i).getUserId() == myUser.getUserId())
+                    groupMembers.remove(i);
             }
         }
 
@@ -123,36 +155,56 @@ public class profileGroupController implements Initializable {
         showGroupData();
     }
 
-    public void deleteMember(){
-
-    }
+    String defaultImagePath = "/com/whisper/client/images/profile.png";
+    Image defaultImage = new Image(getClass().getResourceAsStream(defaultImagePath));
 
     private void showGroupData() {
         groupName.setText(roomChat.getGroupName());
         groupDescription.appendText(roomChat.getDescription());
-//        groupImage.setImage(new Image(new ByteArrayInputStream(roomChat.getPhotoBlob()));
+//        groupImage.setImage(new Image(new ByteArrayInputStream(roomChat.getPhotoBlob().toString().getBytes())));
+        Blob photoBlob = roomChat.getPhotoBlob();
+        if (photoBlob != null) {
+            try {
+                InputStream inputStream = photoBlob.getBinaryStream();
+                Image image = new Image(inputStream);
+                groupImage.setImage(image);
+            } catch (SQLException e) {
+                e.printStackTrace(); // Handle the exception appropriately
+            }
+
+        }else {
+            groupImage.setImage(defaultImage);
+        }
     }
 
     @javafx.fxml.FXML
     public void onClickRemove(ActionEvent actionEvent) {
+//        System.out.println(groupMembers.size());
         int userIndex = groupMembersList.getSelectionModel().getSelectedIndex();
-        HBox removedBox = boxes.remove(userIndex);
-//        System.out.println(userIndex);
-//        System.out.println(groupMembers.size());
-//        System.out.println(groupMembers.size());
+        if (groupMembersList.getSelectionModel().getSelectedIndex() == -1){
+            dialogueManager.showInformationDialog("Button Disabled", "No member selected!");
+            return;
+        }
         chattingService.removeGroupMember(5, groupMembers.get(userIndex).getUserId());
+        System.out.println(groupMembers.size());
+//        contactList.setItems(othersBox);
+        HBox removedBox = boxes.remove(userIndex);
+        otherContacts.add(groupMembers.get(userIndex));
         groupMembers.remove(userIndex);
         othersBox.add(removedBox);
-        contactList.setItems(othersBox);
-        otherContacts.add(groupMembers.get(userIndex));
     }
 
     @javafx.fxml.FXML
     public void onClickAdd(ActionEvent actionEvent) {
         int userIndex = contactList.getSelectionModel().getSelectedIndex();
+        if (contactList.getSelectionModel().getSelectedIndex() == -1){
+            dialogueManager.showInformationDialog("Button Disabled", "No contact selected!");
+            return;
+        }
         HBox box = (HBox) contactList.getSelectionModel().getSelectedItem();
+        chattingService.addGroupMember(5, otherContacts.get(userIndex).getUserId());
+        groupMembers.add(otherContacts.get(userIndex));
         boxes.add(box);
         othersBox.remove(box);
-        chattingService.addGroupMember(5, otherContacts.get(userIndex).getUserId());
     }
 }
